feat: crear un tema nuevo

+++ src/app/registrar/registrar.page.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastController } from '@ionic/angular';
+import {NgForm} from '@angular/forms';
+
 @Component({
   selector: 'app-registrar',
   templateUrl: './registrar.page.html',
@@ -12,29 +14,51 @@ export class RegistrarPage implements OnInit {
   user: string = '';
   email: string = '';
   password: string = '';
+  confirmPassword: string ='';
 
   constructor(private router: Router,private toastController: ToastController) { }
 
   ngOnInit() {
   }
 
-registrar (){
-  localStorage.setItem('userEmail', this.email);
-  localStorage.setItem('userPassword', this.password);
-  localStorage.setItem('userName' , this.user);
+registrar (form: NgForm){
+console.log(form.value);
+console.log(form.valid)
 
-  async presentToast() {
-    const toast = await this.toastController.create({
-      message: 'Credenciales incorrectas',
-      duration: 1500,
-      position: 'bottom',
-      color: 'danger',
-    });
+  if(form.valid){
+    if (form.value.password!=form.value.confirmPassword){
 
-    await toast.present();
+      this.presentToast('Contraseñas diferentes', 'warning');
+     
+    }else{
+      
+      localStorage.setItem('userEmail', this.email);
+      localStorage.setItem('userPassword', this.password);
+      localStorage.setItem('userName' , this.user);  
+      this.presentToast('Registro exitoso', 'primary');
+      this.router.navigate(['/login']); 
+    
+    }
+    
+  }else{
+    this.presentToast('Rellena todos los campos','warning');        
   }
+
+
+ 
+
   
-  this.router.navigate(['/login']);
+}
+
+async presentToast(message: string, color: string,) {
+  const toast = await this.toastController.create({
+    message: message,
+    duration: 1500,
+    position: 'bottom',
+    color: color,
+  });
+
+  await toast.present();
 }
 
 }
